# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Politico is a platform that enables citizens give their mandate to politicians runnning for different government offices while building trust in the process through transparency",
    "version": "1.0.0",
    "title": "Politico",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "amaechichuks2000@yahoo.com"
    },
    "license": {
      "name": "GNU General Public License",
      "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
    },
    "header": {
      "name": "authorization"
    }
  },
  "host": "politico-software.herokuapp.com",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/parties/{id}": {
      "get": {
        "tags": [
          "parties"
        ],
        "summary": "Find Parties by ID",
        "description": "Returns a single political parties",
        "operationId": "getPartyById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Party to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful got politcal party",
            "schema": {
              "$ref": "#/definitions/parties"
            }
          },
          "404": {
            "description": "party not found"
          }
        }
      }
    },
...
    "/parties/{id}/": {
      "delete": {
        "tags": [
          "parties"
        ],
        "summary": "Delete a polical party by id",
        "description": "Delete a politcal party by id",
        "operationId": "DeletePartyById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of political party to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "schema": {
              "$ref": "#/definitions/parties"
            }
          },
          "404": {
            "description": "No party record found"
          },
          "403": {
            "description": "user is not authorized to delete this party."
          }
        }
      }
    },
...
  },
  "definitions": {
    "parties": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "format": "int64"
        },
        "hqAddress": {
          "type": "string"
        },
        "logoUrl": {
          "type": "string",
          "format": "int64"
        },
        "registereAt": {
          "type": "string"
        }
      },
      "xml": {
        "name": "parties"
      }
    },
...
  },
...
}